main.dart を分割しホーム画面と在庫画面を別ファイル化
